Use react client code for rendering svg
